Tpll.Context

Description

Documentation

type Context = Map String ContextValue

A HashMap that holds the context that is passed to the parser

You can generate a context by using the ctx function:

let ctx' = ctx [("foo", cInt 1), ("bar", cStr "2")]

data ContextValue

All possible values that can be passed to the context.

It's also possible to group values by using CAssoc and CList:

CAssoc (cInt 1, cStr "foo")

CList [cInt 1, cStr "foo", cAssoc (cInt 2, cStr "bar")]

Constructors

CStr Bool String

CInt Bool Int

CDouble Bool Double

CAssoc Bool (ContextValue, ContextValue)

CList Bool [ContextValue]

cStr :: String -> ContextValue

cInt :: Int -> ContextValue

cInteger :: Integer -> ContextValue

cDouble :: Double -> ContextValue

cAssoc :: (ContextValue, ContextValue) -> ContextValue

cList :: [ContextValue] -> ContextValue

ctx :: [(String, ContextValue)] -> Context

Helper function to create a Context HashMap from a list.

Examples:

>>> let foo = ctx [("foo", cStr "bar")]
>>> let Just (CStr False r) = lookup "foo" foo
>>> r
"bar"

>>> let bar = ctx [("foo", cStr "bar"), ("bar", cInt 2)]
>>> let Just (CInt False r) = lookup "bar" bar
>>> r
2

resolveCtx :: Context -> String -> Maybe ContextValue

Function used by template tags to resolve values passed as arguments.

Examples:

>>> let ctx' = ctx []
>>> let Just (CStr False r) = resolveCtx ctx' "\"foo\""
>>> r
"foo"

>>> let ctx' = ctx []
>>> let Nothing = resolveCtx ctx' "foo"

>>> let ctx' = ctx [("foo", cInt 42)]
>>> let Just (CInt False r) = resolveCtx ctx' "foo"
>>> r
42

ctxToString :: Maybe ContextValue -> String

Function used to convert a ContextValue into a String.

Examples:

>>> ctxToString (Just (cStr "foo"))
"foo"

>>> ctxToString (Just (cInt 1))
"1"

>>> ctxToString (Just (cInteger 2))
"2"

>>> ctxToString (Just (cDouble 3.14))
"3.14"