Tpll.Tags

Description

Documentation

type Tag = Context -> Token -> [Token] -> TagAction

Defines a template tag that accepts three arguments: Context, Token, and the list of remaining tokens, not consumed by the template tag.

Returns a TagAction.

type Filter = Context -> Maybe ContextValue -> Maybe ContextValue

Defines a template filter that accepts two arguments: Context and ContextValue.

type Tags = (Map String Tag, Map String Filter)

Defines two maps containing template tags (Tag) and filters (Filter).

data TagAction

The action returned by the template tags. It can be either Render or RenderBlock.

The Render action returns a list containing the remaining tokens (those not consumed by the parser) and the string to be rendered. The RenderBlock action returns a context list, the old context, a list containing the remaining tokens and a "final" token.

The context list returned by the RenderBlock action can be used by loop tags, meaning that the block will be iterated over all contexts returned. The "final" token is used to define where the block ends. For example, the final token for the {% for x in list %} tag is the {% endfor %} tag.

Constructors

Render ([Token], IO String)

RenderBlock ([Context], Context, [Token], Token)

tags :: [(String, Tag)] -> [(String, Filter)] -> Tags

Constructs the Tags containing template tags and filters.