Tpll.Tags.DefaultFilters

Description

Documentation

upperFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|upper }}

Transforms text to uppercase.

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cStr, cDouble)
>>> import Tpll.Tags.Default (getAllDefaultTags)
>>> 
>>> let ctx' = ctx [("x", cStr "foo"), ("y", cDouble 3.14)]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ x|upper }}"
"FOO"

>>> parseString ctx' tags' "{{ y|upper }}"
"3.14"

lowerFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|lower }}

Transforms text to lowercase.

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cStr, cDouble)
>>> import Tpll.Tags.Default (getAllDefaultTags)
>>> 
>>> let ctx' = ctx [("x", cStr "Bar"), ("y", cDouble 3.14)]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ x|lower }}"
"bar"

>>> parseString ctx' tags' "{{ y|lower }}"
"3.14"

capFirstFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|capfirst }}

Capitalizes the first character of the value.

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cStr)
>>> import Tpll.Tags.Default (getAllDefaultTags)
>>> 
>>> let ctx' = ctx [("x", cStr "this is a string.")]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ x|capfirst }}"
"This is a string."

>>> parseString ctx' tags' "{{ y|capfirst }}"
""

titleFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|title }}

Capitalizes the first character of each words.

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cStr)
>>> import Tpll.Tags.Default (getAllDefaultTags)
>>> 
>>> let ctx' = ctx [("x", cStr "this is a string.")]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ x|title }}"
"This Is A String."

firstFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|first }}

Returns the first item in a list.

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cInt, cList)
>>> import Tpll.Tags.Default (getAllDefaultTags)
>>> 
>>> let ctx' = ctx [("x", cList [cInt 42, cInt 1, cInt 2])]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ x|first }}"
"42"

safeFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|safe }}

Marks a string as safe for priting HTML.

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cStr)
>>> import Tpll.Tags.Default (getAllDefaultTags)
>>> 
>>> let ctx' = ctx [("x", cStr "<b>foo</b>")]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ x }}"
"&lt;b&gt;foo&lt;/b&gt;"

>>> parseString ctx' tags' "{{ x|safe }}"
"<b>foo</b>"

defaultFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|default:"nothing" }}

If value evaluates to False, uses the given value.

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cStr)
>>> import Tpll.Tags.Default (getAllDefaultTags)
>>> 
>>> let ctx' = ctx [("x", cStr "foo")]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ x|default:\"bar\" }}"
"foo"

>>> parseString ctx' tags' "{{ y|default:\"bar\" }}"
"bar"

dateFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|date:"format" }}

Converts value to the given format.

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cStr, cUTCTime, cIOUTCTime)
>>> import Tpll.Tags.Default (getAllDefaultTags)
>>> 
>>> import Data.Time.Calendar (fromGregorian)
>>> import Data.Time.Clock
>>> 
>>> let ioutctime = return $ UTCTime (fromGregorian 2011 12 16) (fromIntegral $ 12 * 3600)
>>> let utctime = UTCTime (fromGregorian 2011 12 16) (fromIntegral $ 12 * 3600)

>>> let ctx' = ctx [("x", cIOUTCTime ioutctime), ("y", cUTCTime utctime)]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ x|date }}"
"2011-12-16 12:12:00"

>>> parseString ctx' tags' "{{ x|date:\"%Y-%m-%d\" }}"
"2011-12-16"

>>> parseString ctx' tags' "{{ y|date }}"
"2011-12-16 12:12:00"

>>> parseString ctx' tags' "{{ y|date:\"%Y-%m-%d\" }}"
"2011-12-16"

addFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|add:1 }}

Adds the argument to the value

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cInt, cDouble, cStr)
>>> import Tpll.Tags.Default (getAllDefaultTags)
>>> 

>>> let ctx' = ctx [("x", cInt 1), ("y", cStr "foo"), ("z", cDouble 1.0)]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ x|add:1 }}"
"2"

>>> parseString ctx' tags' "{{ y|add:\"bar\" }}"
"foobar"

>>> parseString ctx' tags' "{{ z|add:2.14 }}"
"3.14"

>>> parseString ctx' tags' "{{ z|add:2 }}"
"3.0"

>>> parseString ctx' tags' "{{ y|add:2 }}"
"foo"

lastFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|last }}

Returns the last item in a list.

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cInt, cDouble, cStr, cList, cAssoc)
>>> import Tpll.Tags.Default (getAllDefaultTags)

>>> let ctx' = ctx [("x", cList [cInt 1, cInt 2]), ("y", cStr "foo"), ("z", cAssoc (cInt 3, cInt 4))]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ x|last }}"
"2"

>>> parseString ctx' tags' "{{ y|last }}"
"o"

>>> parseString ctx' tags' "{{ z|last }}"
"4"

>>> parseString ctx' tags' "{{ unknown|last }}"
""

joinFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|join:"separator"" }}

Joins a list of strings.

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cInt, cDouble, cStr, cList, cAssoc)
>>> import Tpll.Tags.Default (getAllDefaultTags)

>>> let ctx' = ctx [("foo", cList [cStr "foo", cStr "bar"])]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ foo|join }}"
"foobar"

>>> parseString ctx' tags' "{{ foo|join:\" -- \" }}"
"foo -- bar"

ljustFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|ljust:10 }}

Left-aligns the value in a field of a given width.

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cStr, cInt)
>>> import Tpll.Tags.Default (getAllDefaultTags)

>>> let ctx' = ctx [("a", cStr "foo"), ("b", cStr "haskell"), ("n", cInt 7)]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ a|ljust:10 }}"
"foo       "

>>> parseString ctx' tags' "{{ b|ljust:n }}"
"haskell"

rjustFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|rjust:10 }}

Right-aligns the value in a field of a given width.

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cStr, cInt)
>>> import Tpll.Tags.Default (getAllDefaultTags)

>>> let ctx' = ctx [("a", cStr "foo"), ("b", cStr "haskell"), ("n", cInt 7)]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ a|rjust:10 }}"
"       foo"

>>> parseString ctx' tags' "{{ b|rjust:n }}"
"haskell"

truncatecharsFilter :: Context -> Maybe ContextValue -> Maybe ContextValue -> Maybe ContextValue

{{ arg|truncatechars:10 }}

Truncates a string if it is longer than the specified number of characters. Truncated strings will end with a translatable ellipsis sequence (”...”).

Examples:

>>> import Tpll.Parser (parseString)
>>> import Tpll.Context (ctx, cStr, cInt)
>>> import Tpll.Tags.Default (getAllDefaultTags)

>>> let ctx' = ctx [("a", cStr "foo"), ("b", cStr "haskell"), ("n", cInt 3)]
>>> let tags' = getAllDefaultTags

>>> parseString ctx' tags' "{{ a|truncatechars:3 }}"
"foo"

>>> parseString ctx' tags' "{{ b|truncatechars:n }}"
"has..."